Fixed getline() name conflict
Up until now, I’ve been using ‘getbigline()’ instead of ‘getline()’ per
K&R due to a name conflict I was getting when compiling on OS X using
CLANG. Turns out this is due to the POSIX 2008 standard incorporating
getline() into the standard.  But, there is a way to undo it so that
these exercises will compile with the original getline() name.